feat: type을 int에서 long으로 변경
- MemoryDb에서 JDBC Db로 변경할 예정이므로 long 타입으로 변경하였다.

+++ src/main/java/codesquad/springcafe/controller/ArticleController.java
@@ -34,7 +34,7 @@ public class ArticleController {
     @PostMapping("/write")
     public String write(@ModelAttribute Article article) {
         article.setCreationTime(new Date());
-        int id = articleService.addArticle(article);
+        int id = (int) articleService.addArticle(article);
         logger.info("[" + id + "번째 게시글 생성 완료] - " + article);
 
         return "redirect:/";
@@ -45,7 +45,7 @@ public class ArticleController {
         Article article = articleService.findArticleById(articleId - 1);
         logger.info("[" + articleId + "번째 게시글 가져오기 성공] - " + article.toString());
 
-        int viewCount = articleService.increaseViewCount(article);
+        long viewCount = articleService.increaseViewCount(article);
         logger.info("[" + articleId + "번째 게시글 조회수 : " + viewCount);
 
         model.addAttribute("article", article);
